+++ src/click.c
@@ -147,15 +147,15 @@ static bool tiling_resize(Con *con, xcb_button_press_event_t *event, const click
         return tiling_resize_for_border(con, BORDER_TOP, event);
     }
 
-    if (event->event_x >= 0 && event->event_x <= bsr.x &&
-        event->event_y >= bsr.y && event->event_y <= con->rect.height + bsr.height)
+    if (event->event_x >= 0 && event->event_x <= (int32_t)bsr.x &&
+        event->event_y >= (int32_t)bsr.y && event->event_y <= (int32_t)(con->rect.height + bsr.height))
         return tiling_resize_for_border(con, BORDER_LEFT, event);
 
-    if (event->event_x >= (con->window_rect.x + con->window_rect.width) &&
-        event->event_y >= bsr.y && event->event_y <= con->rect.height + bsr.height)
+    if (event->event_x >= (int32_t)(con->window_rect.x + con->window_rect.width) &&
+        event->event_y >= (int32_t)bsr.y && event->event_y <= (int32_t)(con->rect.height + bsr.height))
         return tiling_resize_for_border(con, BORDER_RIGHT, event);
 
-    if (event->event_y >= (con->window_rect.y + con->window_rect.height))
+    if (event->event_y >= (int32_t)(con->window_rect.y + con->window_rect.height))
         return tiling_resize_for_border(con, BORDER_BOTTOM, event);
 
     return false;
